requirements.txt
asgiref==3.10.0
Django==5.2.8
django-filter==25.2
django-unfold==0.69.0
djangorestframework==3.16.1
django-cors-headers==4.6.0
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
future==1.0.0
pyzk==0.9
sqlparse==0.5.3
qrcode==8.0
python-barcode==0.15.1
Pillow==11.0.0
